+++ docs/main/changelog.rst
@@ -11,6 +11,7 @@ Develop
 -------
 - Improve wiring with adding importing modules and packages from a string
   ``container.wire(modules=["yourapp.module1"])``.
+- Add container wiring configuration ``wiring_config = containers.WiringConfiguration()``.
 - Update documentation and fix typos.
 
 4.36.2